+++ enumeratum-core/compat/src/test/scala-2.11/enumeratum/values/ValueEnumSpec.scala
@@ -11,33 +11,37 @@ import org.scalatest.{ FunSpec, Matchers }
  */
 class ValueEnumSpec extends FunSpec with Matchers {
 
-  describe("withName") {
+  describe("IntEnum") {
 
-    it("should return entries that match the value") {
-      LibraryItem.withValue(1) shouldBe LibraryItem.Book
-      LibraryItem.withValue(2) shouldBe LibraryItem.Movie
-      LibraryItem.withValue(3) shouldBe LibraryItem.Magazine
-      LibraryItem.withValue(4) shouldBe LibraryItem.CD
-    }
+    describe("withName") {
 
-    it("should throw on values that don't map to any entries") {
-      intercept[NoSuchElementException] {
-        LibraryItem.withValue(5)
+      it("should return entries that match the value") {
+        LibraryItem.withValue(1) shouldBe LibraryItem.Book
+        LibraryItem.withValue(2) shouldBe LibraryItem.Movie
+        LibraryItem.withValue(3) shouldBe LibraryItem.Magazine
+        LibraryItem.withValue(4) shouldBe LibraryItem.CD
       }
+
+      it("should throw on values that don't map to any entries") {
+        intercept[NoSuchElementException] {
+          LibraryItem.withValue(5)
+        }
+      }
+
     }
 
-  }
+    describe("withValueOpt") {
 
-  describe("withValueOpt") {
+      it("should return Some(entry) that match the value") {
+        LibraryItem.withValueOpt(1) shouldBe Some(LibraryItem.Book)
+        LibraryItem.withValueOpt(2) shouldBe Some(LibraryItem.Movie)
+        LibraryItem.withValueOpt(3) shouldBe Some(LibraryItem.Magazine)
+        LibraryItem.withValueOpt(4) shouldBe Some(LibraryItem.CD)
+      }
+      it("should return None when given values that do not map to any entries") {
+        LibraryItem.withValueOpt(5) shouldBe None
+      }
 
-    it("should return Some(entry) that match the value") {
-      LibraryItem.withValueOpt(1) shouldBe Some(LibraryItem.Book)
-      LibraryItem.withValueOpt(2) shouldBe Some(LibraryItem.Movie)
-      LibraryItem.withValueOpt(3) shouldBe Some(LibraryItem.Magazine)
-      LibraryItem.withValueOpt(4) shouldBe Some(LibraryItem.CD)
-    }
-    it("should return None when given values that do not map to any entries") {
-      LibraryItem.withValueOpt(5) shouldBe None
     }
 
   }
