+++ macros/compat/src/main/scala-2.10/enumeratum/ContextUtils.scala
@@ -1,0 +1,14 @@
+package enumeratum
+
+object ContextUtils {
+
+  type Context = scala.reflect.macros.Context
+
+  /**
+   * Returns a TermName
+   */
+  def termName(c: Context)(name: String) = {
+    c.universe.newTermName(name)
+  }
+
+}
