+++ macros/compat/src/main/scala-2.11/enumeratum/ContextUtils.scala
@@ -1,0 +1,14 @@
+package enumeratum
+
+object ContextUtils {
+
+  type Context = scala.reflect.macros.blackbox.Context
+
+  /**
+   * Returns a TermName
+   */
+  def termName(c: Context)(name: String) = {
+    c.universe.TermName(name)
+  }
+
+}
